chore: move test files to the __tests__ directory
