+++ src/components/__tests__/ErrorBoundary.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 
-import ErrorBoundary from './ErrorBoundary';
+import ErrorBoundary from '@components/ErrorBoundary';
 
 import type { JSX } from 'react';
